+++ src/main/java/mainProgram/services/JobService.java
@@ -17,33 +17,26 @@ public class JobService implements BaseSearchService<Job> {
     private final JobPartRepository jobPartRepository;
     private final ProductRepository productRepository;
 
-    public JobService(
-            JobRepository jobRepository,
-            JobPartRepository jobPartRepository,
-            ProductRepository productRepository
-    ) {
+    public JobService(JobRepository jobRepository, JobPartRepository jobPartRepository, ProductRepository productRepository) {
         this.jobRepository = jobRepository;
         this.jobPartRepository = jobPartRepository;
         this.productRepository = productRepository;
     }
 
-    public Job getJobById(Long id) {
-        return jobRepository
-                .findById(Math.toIntExact(id))
-                .orElseThrow(() -> new RuntimeException("Job not found"));
+    public Job getJobById(int id) {
+        return jobRepository.findById(Math.toIntExact(id)).orElseThrow(() -> new RuntimeException("Job not found"));
     }
 
     /// Find the parts associated with a repair
-    public List<JobPart> getPartsForJob(Long jobId) {
+    public List<JobPart> getPartsForJob(int jobId) {
         return jobPartRepository.findByJobId(jobId);
     }
 
+
     /// Add a new product to a repair, using the JobPart join-table
-    public void addProductToRepair(Long repairId, Long productId, int quantity) {
+    public void addProductToRepair(int repairId, int productId, int quantity) {
         Job repair = getJobById(repairId);
-        Product product = productRepository
-                .findById(productId)
-                .orElseThrow(() -> new RuntimeException("Product not found"));
+        Product product = productRepository.findById(productId).orElseThrow(() -> new RuntimeException("Product not found"));
 
         // Create a new job-part object and add it to the join-table
         JobPart jobpart = new JobPart(repair, product, quantity);
